refactor to support RN 0.40
