+++ android/src/main/java/com/jumio/react/JumioModule.java
@@ -5,11 +5,18 @@
 
 package com.jumio.react;
 
+import android.app.Activity;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.support.v4.app.ActivityCompat;
 import android.util.Log;
+import android.widget.Toast;
+
 import java.util.ArrayList;
 
+import com.facebook.react.bridge.ActivityEventListener;
 import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 
@@ -26,9 +33,11 @@ import com.jumio.dv.DocumentVerificationSDK;
 import com.jumio.nv.*;
 import com.jumio.nv.data.document.NVDocumentType;
 import com.jumio.nv.data.document.NVDocumentVariant;
-import com.jumio.sdk.SDKExpiredException;
 
-public class JumioModule extends ReactContextBaseJavaModule {
+import org.json.JSONException;
+import org.json.JSONObject;
+
+public class JumioModule extends ReactContextBaseJavaModule implements ActivityEventListener {
     
     private final static String TAG = "JumioMobileSDK";
     public static final int PERMISSION_REQUEST_CODE_BAM = 300;
@@ -47,7 +56,124 @@ public class JumioModule extends ReactContextBaseJavaModule {
     public String getName() {
         return "JumioMobileSDK";
     }
-    
+
+    protected static void onRequestPermissionsResult(ReactContext reactContext, int requestCode, String[] permissions, int[] grantResults) {
+        boolean allGranted = true;
+        for (int grantResult : grantResults) {
+            if (grantResult != PackageManager.PERMISSION_GRANTED) {
+                allGranted = false;
+                break;
+            }
+        }
+
+        if (allGranted) {
+            if (requestCode == JumioModule.PERMISSION_REQUEST_CODE_BAM) {
+                startSdk(reactContext, JumioModule.bamSDK);
+            } else if (requestCode == JumioModule.PERMISSION_REQUEST_CODE_NETVERIFY) {
+                startSdk(reactContext, JumioModule.netverifySDK);
+            } else if (requestCode == JumioModule.PERMISSION_REQUEST_CODE_DOCUMENT_VERIFICATION) {
+                startSdk(reactContext, JumioModule.documentVerificationSDK);
+            }
+        } else {
+            Toast.makeText(reactContext, "You need to grant all required permissions to start the Jumio SDK", Toast.LENGTH_LONG).show();
+        }
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == BamSDK.REQUEST_CODE) {
+            if (resultCode == Activity.RESULT_OK) {
+                BamCardInformation cardInformation = data.getParcelableExtra(BamSDK.EXTRA_CARD_INFORMATION);
+                JSONObject result = new JSONObject();
+                try {
+                    result.put("cardType", cardInformation.getCardType());
+                    result.put("cardNumber", String.valueOf(cardInformation.getCardNumber()));
+                    result.put("cardNumberGrouped", String.valueOf(cardInformation.getCardNumberGrouped()));
+                    result.put("cardNumberMasked", String.valueOf(cardInformation.getCardNumberMasked()));
+                    result.put("cardExpiryMonth", String.valueOf(cardInformation.getCardExpiryDateMonth()));
+                    result.put("cardExpiryYear", String.valueOf(cardInformation.getCardExpiryDateYear()));
+                    result.put("cardExpiryDate", String.valueOf(cardInformation.getCardExpiryDateYear()));
+                    result.put("cardCVV", String.valueOf(cardInformation.getCardCvvCode()));
+                    result.put("cardHolderName", String.valueOf(cardInformation.getCardHolderName()));
+                    result.put("cardSortCode", String.valueOf(cardInformation.getCardSortCode()));
+                    result.put("cardAccountNumber", String.valueOf(cardInformation.getCardAccountNumber()));
+                    result.put("cardSortCodeValid", cardInformation.isCardSortCodeValid());
+                    result.put("cardAccountNumberValid", cardInformation.isCardAccountNumberValid());
+
+                    sendEvent("EventCardInformation", result);
+                    cardInformation.clear();
+                } catch (JSONException e) {
+                    showErrorMessage("Result could not be sent. Try again.");
+                }
+            } else if (resultCode == Activity.RESULT_CANCELED) {
+                String errorMessage = data.getStringExtra(BamSDK.EXTRA_ERROR_MESSAGE);
+                showErrorMessage(errorMessage);
+            }
+        } else if (requestCode == NetverifySDK.REQUEST_CODE) {
+            if (resultCode == Activity.RESULT_OK) {
+                NetverifyDocumentData documentData = (NetverifyDocumentData) data.getParcelableExtra(NetverifySDK.EXTRA_SCAN_DATA);
+                JSONObject result = new JSONObject();
+                try {
+                    result.put("selectedCountry", documentData.getSelectedCountry());
+                    result.put("selectedDocumentType", documentData.getSelectedDocumentType());
+                    result.put("idNumber", documentData.getIdNumber());
+                    result.put("personalNumber", documentData.getPersonalNumber());
+                    result.put("issuingDate", documentData.getIssuingDate());
+                    result.put("expiryDate", documentData.getExpiryDate());
+                    result.put("issuingCountry", documentData.getIssuingCountry());
+                    result.put("lastName", documentData.getLastName());
+                    result.put("firstName", documentData.getFirstName());
+                    result.put("middleName", documentData.getMiddleName());
+                    result.put("dob", documentData.getDob());
+                    result.put("gender", documentData.getGender());
+                    result.put("originatingCountry", documentData.getOriginatingCountry());
+                    result.put("addressLine", documentData.getAddressLine());
+                    result.put("city", documentData.getCity());
+                    result.put("subdivision", documentData.getSubdivision());
+                    result.put("postCode", documentData.getPostCode());
+                    result.put("optionalData1", documentData.getOptionalData1());
+                    result.put("optionalData2", documentData.getOptionalData2());
+                    result.put("placeOfBirth", documentData.getPlaceOfBirth());
+                    result.put("extractionMethod", documentData.getExtractionMethod());
+
+                    // MRZ data if available
+                    if (documentData.getMrzData() != null) {
+                        JSONObject mrzData = new JSONObject();
+                        mrzData.put("format", documentData.getMrzData().getFormat());
+                        mrzData.put("line1", documentData.getMrzData().getMrzLine1());
+                        mrzData.put("line2", documentData.getMrzData().getMrzLine2());
+                        mrzData.put("line3", documentData.getMrzData().getMrzLine3());
+                        mrzData.put("idNumberValid", documentData.getMrzData().idNumberValid());
+                        mrzData.put("dobValid", documentData.getMrzData().dobValid());
+                        mrzData.put("expiryDateValid", documentData.getMrzData().expiryDateValid());
+                        mrzData.put("personalNumberValid", documentData.getMrzData().personalNumberValid());
+                        mrzData.put("compositeValid", documentData.getMrzData().compositeValid());
+                        result.put("mrzData", mrzData);
+                    }
+
+                    // EMRTD data if available
+                    if (documentData.getEMRTDStatus() != null) {
+                        result.put("emrtdStatus", documentData.getEMRTDStatus());
+                    }
+
+                    sendEvent("EventDocumentData", result);
+                } catch (JSONException e) {
+                    showErrorMessage("Result could not be sent. Try again.");
+                }
+            } else if (resultCode == Activity.RESULT_CANCELED) {
+                String errorMessage = data.getStringExtra(NetverifySDK.EXTRA_ERROR_MESSAGE);
+                showErrorMessage(errorMessage);
+            }
+        } else if (requestCode == DocumentVerificationSDK.REQUEST_CODE) {
+            if (resultCode == Activity.RESULT_OK) {
+                sendEvent("EventDocumentVerification", "Document-Verification finished successfully.");
+            } else if (resultCode == Activity.RESULT_CANCELED) {
+                String errorMessage = data.getStringExtra(DocumentVerificationSDK.EXTRA_ERROR_MESSAGE);
+                showErrorMessage(errorMessage);
+            }
+        }
+    }
+
     // Netverify
     
     @ReactMethod
@@ -347,7 +473,7 @@ public class JumioModule extends ReactContextBaseJavaModule {
                 return;
             }
             
-            ActivityCompat.requestPermissions(getReactApplicationContext().getCurrentActivity(), mp, code);
+            ActivityCompat.requestPermissions(getCurrentActivity(), mp, code);
             //The result is received in MainActivity::onRequestPermissionsResult.
         } else {
             startSdk(sdk);
@@ -356,16 +482,37 @@ public class JumioModule extends ReactContextBaseJavaModule {
     
     // Helper methods
     
-    private void startSdk(MobileSDK sdk) {
+    private static void startSdk(ReactContext reactContext, MobileSDK sdk) {
         try {
             sdk.start();
         } catch (MissingPermissionException e) {
-            showErrorMessage(e.getLocalizedMessage());
+            showErrorMessage(reactContext, e.getLocalizedMessage());
         }
     }
-    
-    private void showErrorMessage(String msg) {
+
+    private void startSdk(MobileSDK sdk) {
+        startSdk(getReactApplicationContext(), sdk);
+    }
+
+    private static void showErrorMessage(ReactContext reactContext, String msg) {
         Log.e("Error", msg);
-        getReactApplicationContext().getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class).emit("EventError", msg);
+        reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class).emit("EventError", msg);
+    }
+
+    private void showErrorMessage(String msg) {
+        showErrorMessage(getReactApplicationContext(), msg);
+    }
+
+    private static void sendEvent(ReactContext reactContext, String eventName, String data) {
+        reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                .emit(eventName, data);
+    }
+
+    private void sendEvent(String eventName, JSONObject params) {
+        sendEvent(getReactApplicationContext(), eventName, params.toString());
+    }
+
+    private void sendEvent(String eventName, String msg) {
+        sendEvent(getReactApplicationContext(), eventName, msg);
     }
 }
