+++ android/src/main/java/com/jumio/react/JumioPackage.java
@@ -9,6 +9,7 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.JavaScriptModule;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.uimanager.ViewManager;
 
 import java.util.ArrayList;
@@ -16,6 +17,10 @@ import java.util.Collections;
 import java.util.List;
 
 public class JumioPackage implements ReactPackage {
+    @Override
+    public List<Class<? extends JavaScriptModule>> createJSModules() {
+        return Collections.emptyList();
+    }
     
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
@@ -27,5 +32,9 @@ public class JumioPackage implements ReactPackage {
         List<NativeModule> modules = new ArrayList<>();
         modules.add(new JumioModule(reactContext));
         return modules;
+    }
+
+    public static void onRequestPermissionsResult(ReactContext reactContext, int requestCode, String[] permissions, int[] grantResults) {
+        JumioModule.onRequestPermissionsResult(reactContext, requestCode, permissions, grantResults);
     }
 }
